pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-Warehouse_env"
version = "0.1.0"
description = "Warehouse Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m Warehouse_env.server.app
server = "Warehouse_env.server.app:main"
warehouse-baseline = "Warehouse_env.inference:main"

[tool.setuptools]
include-package-data = true
packages = ["Warehouse_env", "Warehouse_env.server"]
package-dir = { "Warehouse_env" = ".", "Warehouse_env.server" = "server" }